Fix #112590: Split edges crash with loose vertices
A mistake in the backport of be2df46afe58a662b2d0 caused a crash when
there are split edges in some cases. That was caused by changing from a
lambda in the 4.0 version of `IndexMask::foreach_segment` and the
regular iteration over the mask in 3.6. A return in 4.0 is equivalent to
a continue at the top level of the loop in 3.6.