3.6 LTS: Fix crash converting spaces to tabs
The cursor & selection weren't updated after converting between tabs
& spaces, meaning they could be invalid (in the middle of a UTF8 byte
sequence or out of bounds).

Resolve by moving the cursor & selection to the line start.

Note that this was fixed in `main` [0] with support for storing/restoring
the cursor column. This required the addition of tab support for
column/offset conversion which is too large a change for the LTS.

[0]: 66028c8c973dc6a1905d7f4a7339f533383d9ca5

Pull Request: https://projects.blender.org/blender/blender/pulls/112492